requirements.txt
phi
groq
python-dotenv
sqlalchemy
